extract Emit method
